Unix shell
Description- 
This program is a simulation of a UNIX Shell, it takes in commands and arguments from the user to perform and execute specific tasks. The program run a continuously loops to get user input, It then
determines whether this is a shell task or a user task (such as showing history or terminating. A shell
task will fork a child process and invoke execcpv().
If the command and arguments are not followed by an ampersand ‘&’. Then the parent process will wait for the newly created child process to finish before continuing its run. 
If the command and arguments are followed by an ampersand ‘&’ then the child and parent will run concurrently.
The program saves every entry from the user in a history buffer. It will hold a total of 10 previous commands before overwriting the oldest commands.  The max for the history buffer can be changed by editing the H_MAX variable. 
To execute a command
Enter command's name followed by arguments. 
To exit: type “exit”
To view history, type “history”.
To get and execute the most recent command type “!!”
-The shell will get the last command, add it to history again and execute it.