Chat client
Receive services from the server and respond accordingly.

Automatically connected to server when program starts.
The server will then request a username to be used for the session. username must be unused or server will request again

User is shown the menu  
#1. List Users
#2. Chat
 #3. Exit

Entering 1 shows all users connected to server.
Entering 3 will disconnect user from server
Entering 2 will show the user a list of available users to chat with. The user then enters to name they wish to chat with.